+++ src/main/java/com/wardk/meeteam_backend/domain/project/service/ProjectService.java
@@ -1,7 +1,7 @@
 
 package com.wardk.meeteam_backend.domain.project.service;
 
-import com.wardk.meeteam_backend.global.auth.dto.CustomSecurityUserDetails;
+import com.wardk.meeteam_backend.web.auth.dto.CustomSecurityUserDetails;
 import com.wardk.meeteam_backend.web.mainpage.dto.CategoryCondition;
 import com.wardk.meeteam_backend.web.mainpage.dto.ProjectConditionMainPageResponse;
 import com.wardk.meeteam_backend.web.project.dto.*;
@@ -9,7 +9,6 @@ import com.wardk.meeteam_backend.web.projectLike.dto.ProjectWithLikeDto;
 import com.wardk.meeteam_backend.web.projectMember.dto.ProjectUpdateResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
