+++ src/main/java/com/wardk/meeteam_backend/domain/projectMember/service/ProjectApplicationService.java
@@ -1,6 +1,6 @@
 package com.wardk.meeteam_backend.domain.projectMember.service;
 
-import com.wardk.meeteam_backend.global.auth.dto.CustomSecurityUserDetails;
+import com.wardk.meeteam_backend.web.auth.dto.CustomSecurityUserDetails;
 import com.wardk.meeteam_backend.web.projectMember.dto.*;
 
 import java.util.List;
