+++ src/main/java/com/wardk/meeteam_backend/web/auth/dto/register/SkillDto.java
@@ -1,4 +1,4 @@
-package com.wardk.meeteam_backend.global.auth.dto.register;
+package com.wardk.meeteam_backend.web.auth.dto.register;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
